refactor: extract ImmichImportJob and deduplicate walk_dir
Move ImmichImportJob (190 lines) from providers/immich.rs to its own
file at immich_importer.rs. Make walk_dir and collect_candidates public
in importer.rs so both local and Immich importers share the same
directory walking logic instead of duplicating it.

immich.rs: 691 → 487 lines.

Closes #156
